requirements.txt
asgiref==3.5.2
Django==4.1.2
django-decorator-include==3.0
django-extensions==3.2.1
names==0.3.0
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
xmltodict==0.13.0
